fix of app.js and formulaire.js
