+++ client/formulaire.html
@@ -82,7 +82,7 @@
   </section>
   
 
-  <script type="module" src="./js/app.js"></script>
+  <script type="module" src="./js/components/formulaire.js"></script>
 </body>
 
 </html>
